requirements_ecommerce.txt
Flask>=2.3.0
mysql-connector-python>=8.1.0
bcrypt>=4.0.1
websockets>=11.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
